% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landgen.r
\name{landgen}
\alias{landgen}
\title{Simulate a landscape consisting of 2 patch classes (1, 2) superimposed on
a matrix (0). The area remaining as "matrix" (class 0) will naturally be
\code{area} - \code{prop1 x area + prop2 x area}. Patch number (n1, n2)
controls the granularity of the landscape. Note, though, that patches can
become contiguous, so there may not be n discrete patches in the output.}
\usage{
landgen(area = 10000, prop1, prop2, n1, n2)
}
\arguments{
\item{area}{the area of the simulated landscape in m2. By default, it is set
to 10,000 (100 x 100) to simulate a 1 hectare landscape at a resolution of
1 x 1 m.}

\item{prop1}{proportion of total landscape covered by patch class 1}

\item{prop2}{proportion of total landscape covered by patch class 2}

\item{n1}{number of class1 patches}

\item{n2}{number of class2 patches}
}
\value{
A raster
}
\description{
Simulate a landscape consisting of 2 patch classes (1, 2) superimposed on
a matrix (0). The area remaining as "matrix" (class 0) will naturally be
\code{area} - \code{prop1 x area + prop2 x area}. Patch number (n1, n2)
controls the granularity of the landscape. Note, though, that patches can
become contiguous, so there may not be n discrete patches in the output.
}
\examples{
landgen(10000, prop1 = 0.5, prop2 = 0.25, n1 = 10, n2 = 5)
}
