% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landpheno.r
\name{landpheno}
\alias{landpheno}
\title{Extend a floral landscape over a time series, apply a phenology function
to each floral species, and generate a vector of nectar availability through
time.}
\usage{
landpheno(x, first = 0, last = 365)
}
\arguments{
\item{x}{the data frame output of \code{landtab}}

\item{first}{start of time series, by default 0 (= Jan 1)}

\item{last}{end of time series, by default 365 (= Dec 31)}
}
\value{
A data frame including a vector of nectar availability
}
\description{
Extend a floral landscape over a time series, apply a phenology function
to each floral species, and generate a vector of nectar availability through
time.
}
\examples{
landtally(x, first = 40, last = 300)
}
