% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtab.r
\name{landtab}
\alias{landtab}
\title{Tabulate a simulated landscape, populate patch classes with plants, and
add floral data from a trait table.}
\usage{
landtab(x, comp0, comp1, comp2, prob0, prob1, prob2, traitTab, plantNames)
}
\arguments{
\item{x}{the raster output of \code{landgen}}

\item{comp0}{a vector of plant names, corresponding to names in the floral
trait table, that will make up patch class 0; can be length 1 for
monospecific class}

\item{comp1}{a vector of plant names, corresponding to names in the floral
trait table, that will make up patch class 1; can be length 1 for
monospecific class}

\item{comp2}{a vector of plant names, corresponding to names in the floral
trait table, that will make up patch class 2; can be length 1 for
monospecific class}

\item{prob0}{a vector of weights for each plant in comp0, controlling the
relative abundance of each plant within class0; same length as comp0}

\item{prob1}{a vector of weights for each plant in comp1, controlling the
relative abundance of each plant within class1; same length as comp1}

\item{prob2}{a vector of weights for each plant in comp2, controlling the
relative abundance of each plant within class2; same length as comp2}

\item{traitTab}{a table of floral taxa and traits; must at minimum contain
(1) plant name, (2) peak floral density, (3) first bloom, (4) peak bloom,
(5) last bloom, and (6) nectar per flower; ideally also contains relevant
functional traits in the form of categorical variables.}

\item{plantNames}{the name of the column in the trait table that contains the
plant names}
}
\value{
A data frame
}
\description{
Tabulate a simulated landscape, populate patch classes with plants, and
add floral data from a trait table.
}
\examples{
landtab(land1,
    comp0 = c("Cirsium vulgare", "Trifolium repens", "Taraxacum agg."),
    comp1 = c("Brassica napus"),
    comp2 = c("Acer pseudoplatanus", "Crataegus monogyna", "Salix cinerea"),
    prob0 = c(1, 1, 1),
    prob1 = c(1),
    prob2 = c(0.8, 0.15, 0.05),
    traitTab = baude_traits,
    plantNames = "name.latin")
}
