% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.r
\name{pheno}
\alias{pheno}
\title{Model floral phenology as a triangular function of first bloom, last bloom,
and peak bloom; called internally by \code{landpheno}.}
\usage{
pheno(x, start, end, peak)
}
\arguments{
\item{x}{a vector of time steps}

\item{start}{time step of first bloom}

\item{end}{time step of last bloom}

\item{peak}{time step of peak bloom}
}
\value{
a vector of proportional floral density
}
\description{
Model floral phenology as a triangular function of first bloom, last bloom,
and peak bloom; called internally by \code{landpheno}.
}
