% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtally.r
\name{landtally}
\alias{landtally}
\title{Sum up nectar availability by various criteria (e.g. morphology, patch class,
floral taxon).}
\usage{
landtally(x, criterion = NULL)
}
\arguments{
\item{x}{the data frame output of \code{landpheno}}

\item{criterion}{a categorical variable in the input data frame \code{x}
by which nectar availability will be parsed}
}
\value{
A data frame of summed nectar availability
}
\description{
Sum up nectar availability by various criteria (e.g. morphology, patch class,
floral taxon).
}
